+++ tests/__quickie_test/__init__.py
@@ -2,7 +2,7 @@ from quickie import tasks
 
 from . import nested
 
-QCK_NAMESPACES = {
+NAMESPACES = {
     "nested": nested,
 }